Fixed bugs and added image sizes for Retina.
Fixed bugs in content-excerpt.php and functions.php.
Added image sizes for high resolution displays to Media library.
Inserting images with these new sizes, attributes of width and height
in image tags are set to half.
Added a taxonomy for high resolution displays to Media library to be
used  in image.php.